MOVE: pace class to correct package
